Fix check for retransmission of discarded block segments (#546)
We are waiting for the sequence number to be the same as the last good
known sequence number, to start considering the messages.  However,
this is wrong because the SDO server will start sending the discarded
segments at the start of a new block, which means the sequence number
restarts from 1.

Reset the internal _ackseq counter to zero after each acknowledgement,
which results in the correct value being compared after
retransmission.

Add a block retransmit test. This test passes with this implementation
but will fail with an invalid CRC without fix for discarded segments.

Co-authored-by: André Colomb <src@andre.colomb.de>