Gracefully handle readonly entries in PdoMap.save() (fixes #541) (#593)
* Skip saving and log read-only PDO parameters.

For PDO communication or mapping parameters which are read-only on a
node, the PdoMap.save() method currently aborts with an SDO exception.
However, individual parameters may simply not support configuration
changes.  Rely on the Object Dictionary to tell which ones are
expected to fail and skip them.  For communication record entries, log
what is not being saved.

* Switch to a LocalNode in TestPDO to not require a network.

* Test PdoMap.save() skipping readonly entries.

Co-authored-by: Erlend E. Aasland <erlend@python.org>