build: Bump minimum Python version to 3.9 (fixes #585) (#610)
* Update license metadata.

Avoid deprecation warnings with modern setuptools, which will stop
supporting a table as the project.license option in pyproject.toml
soon.  Require a modern version >= 77 to make use of the alternative
options.

* Modernize Dict and List type annotations.

Since Python 3.9 we can use the built-in types directly.

* Simplify type import.

* Modernize Iterator type annotations.

Since Python 3.9 the collections.abc module can be used directly for
typing.