Disregard OD variable description in SdoClient.upload() (#592)
* Disregard OD variable description in SdoClient.upload().

The upload method should behave as a raw producer of byte data.
Interpretation of the returned data based on the Object Dictionary is
the responsibility of the SdoVariable class, which delegates to the
generic ODVariable decoding functions.

Move the data truncation to the method SdoVariable.get_data(), where
access to the ODVariable is certain.

The only truncation that still happens is based on the response size
specified by the server, which might be smaller for e.g. expedited
upload.  Extend the upload() function docstring to clarify the changed
behavior.  Adjust the test case for expedited upload with unspecified
size in the response, which is the only incompatible change.

* Test expedited upload with 8 byte padded frame.

The existing test response is actually wrong, since the standard
mandates four bytes "data" regardless of the other flags.